WIP: Add "fit" method. Add executor abstraction.
Also, shuffle a bunch of definitions around. (There's a bunch of
refactoring stuffed into this commit.)
